% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bang.r
\name{bang}
\alias{bang}
\title{Uses the logical not}
\usage{
bang(x)
}
\arguments{
\item{x}{a logical object}
}
\value{
a logical object
}
\description{
A wrapper function for \code{!}.
}
\examples{
# load the data
data(nhanes)

# check whether values of smq020 are not missing
nhanes$smq020 |> is.na() |> bang()

# check whether values of smq020 are not missing
nhanes |> dollar(smq020) |> is.na() |> bang()
}
