% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.r
\name{bracket}
\alias{bracket}
\title{Subsets a data using bracket notation}
\usage{
bracket(data_object, index)
}
\arguments{
\item{data_object}{a data object}

\item{index}{a character, logical, or numeric vector}
}
\value{
a subset of the data
}
\description{
A wrapper function for bracket notation. The function only allows one index; for two-dimensional objects, the index is used on the columns.
}
\examples{
# load package
data(nhanes)

# look at the first few values of dmdborn4
nhanes |> bracket('dmdborn4') |> head()
}
