% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dollar.r
\name{dollar}
\alias{dollar}
\title{Selects a portion of a data object using dollar-sign selection}
\usage{
dollar(data_object, variable_name)
}
\arguments{
\item{data_object}{the data object}

\item{variable_name}{the variable name or element name}
}
\value{
a portion of the data object
}
\description{
A wrapper function for dollar-sign notation.
}
\examples{
# load data
data(nhanes)

# summarize lbxtc
nhanes |> dollar(lbxtc) |> summary()
}
