% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.r
\name{et}
\alias{et}
\title{Tables variables}
\usage{
et(data_frame, variable_name)
}
\arguments{
\item{data_frame}{the data.frame}

\item{variable_name}{the variable name}
}
\value{
a data frame
}
\description{
Creates a table for a variable, mimicking \code{count()} (from 'dplyr'). ET stands for exploratory table.
}
\examples{
# load data
data(nhanes)

# define country
nhanes<-nhanes |> transform(
  country=base_match(nhanes$dmdborn4,'USA'=1,'Other'=2)
)

# examine variable, with native piping
nhanes |> et(country)

# examine variable, with dollar-sign notation
et(nhanes$country)
}
