% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes.r
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{National Health and Nutrition Examination Survey data}
\format{
\subsection{\code{nhanes}}{

A data frame with 8,153 rows and 3 columns:
\describe{
\item{seqn}{a unique identifier}
\item{dmdborn4}{country of birth}
\item{lbxtc}{total cholesterol}
}
}
}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx?Cycle=2021-2023}
}
\usage{
nhanes
}
\description{
A subset of data from the National Health and Nutrition Examination Survey
}
\keyword{datasets}
