% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.r
\name{not}
\alias{not}
\title{Uses the logical not}
\usage{
not(x)
}
\arguments{
\item{x}{a logical object}
}
\value{
a logical object
}
\description{
A wrapper function for \code{!}. This is the same thing as \code{bang()}.
}
\examples{
# load the data
data(nhanes)

# check whether values of smq020 are not missing
nhanes$smq020 |> is.na() |> not()

# check whether values of smq020 are not missing
nhanes |> dollar(smq020) |> is.na() |> not()
}
