% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_cor.r
\name{p_cor}
\alias{p_cor}
\title{Finds correlation coefficients, with support for piping}
\usage{
p_cor(data, x, y, ...)
}
\arguments{
\item{data}{the data}

\item{x}{one of the two variables}

\item{y}{one of the two variables}

\item{...}{additional arguments passed to \code{\link[stats:cor]{stats::cor()}}.}
}
\value{
a numeric object
}
\description{
A wrapper function to \code{cor()}, with support for piping.
}
\examples{
# load the data
data(nhanes)

# get correlation coefficient for systolic and diastolic
nhanes |> p_cor(bpxosy1,bpxodi1,use='complete.obs')
}
