% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_t.test.r
\name{p_t.test}
\alias{p_t.test}
\title{Conducts t-tests, with support for piping}
\usage{
p_t.test(data, x = NULL, y = NULL, formula = NULL, ...)
}
\arguments{
\item{data}{the data}

\item{x}{one of two variables}

\item{y}{one of two variables}

\item{formula}{a formula}

\item{...}{additional arguments passed to \code{\link[stats:t.test]{stats::t.test()}}.}
}
\value{
an \code{htest} object
}
\description{
A wrapper function to \code{t.test()}, with the \code{data} argument listed first to support piping.
}
\examples{
# load the data
data(nhanes)

# define smoking
nhanes<-nhanes |> transform(
  smoking=base_match(smq020,'No'=2,'History of smoking'=1)
)

# conduct a one-sample t-test
nhanes |> p_t.test(bpxosy1)

# conduct a two-sample t-test, using formula notation
nhanes |> p_t.test(bpxosy1~smoking)

# conduct a two-sample t-test, using formula notation
nhanes |> p_t.test(formula=bpxosy1~smoking)

# conduct a paired t-test, using x and y
nhanes |> p_t.test(bpxosy1,bpxosy2,paired=TRUE)
}
