% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pglm.r
\name{pglm}
\alias{pglm}
\title{Fits generalized linear models, with support for piping}
\usage{
pglm(data, formula, ...)
}
\arguments{
\item{data}{the data}

\item{formula}{the formula}

\item{...}{additional arguments passed to \code{\link[stats:glm]{stats::glm()}}.}
}
\value{
a \code{glm} object
}
\description{
A wrapper function to \code{glm()}, with the \code{data} argument listed first to enable piping.
}
\examples{
# load the data
data(mtcars)

# fit a model
mtcars |> pglm(am~wt+qsec) |> summary()
}
