% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm.r
\name{plm}
\alias{plm}
\title{\code{plm.r()}}
\usage{
plm(data, formula, ...)
}
\arguments{
\item{data}{the data}

\item{formula}{the formula}

\item{...}{additional arguments passed to \code{\link[stats:lm]{stats::lm()}}.}
}
\value{
an \code{lm} object
}
\description{
A wrapper function to \code{lm()}, with the \code{data} argument listed first to enable piping.
}
\examples{
# load the data
data(mtcars)

mtcars |> plm(drat~wt+qsec) |> summary()
}
