% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bang.r
\name{bang}
\alias{bang}
\title{Flips a logical object.}
\usage{
bang(x)
}
\arguments{
\item{x}{a logical object}
}
\value{
a logical object
}
\description{
A wrapper function to \code{!}.
}
\examples{
# load the data
data(nhanes)

# get correlation coefficient for systolic and diastolic
nhanes$smq020 |> is.na() |> bang()
}
