% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not.r
\name{not}
\alias{not}
\title{Flips a logical object.}
\usage{
not(x)
}
\arguments{
\item{x}{a logical object}
}
\value{
a numeric object
}
\description{
A wrapper function to \code{!}.
}
\examples{
# load the data
data(nhanes)

# get correlation coefficient for systolic and diastolic
nhanes$smq020 |> is.na() |> not()
}
